+++ src/lightning_fabric/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [1.9.6] - 2023-04-DD
+
+### Fixed
+
+~nothing~
+
+
 ## [1.9.5] - 2023-03-30
 
 ### Changed